refactor(design, foundation-demo): move some classes in foundation-demo into design with attribute components
move some classes in foundation-demo into design with attribute components
