+++ apps/foundation-demo/src/app/core/image-gallery/components/image-gallery.component.html
@@ -3,9 +3,9 @@
     <div class="image-gallery__gallery-image-wrapper" *ngFor="let image of images">
       <img 
         alt="{{image.label}}"
-        src="{{image.url}}"  
-        class="image-gallery__gallery-image"
-        [class.image-gallery__gallery-image--selected]="image.url === (selectedImage$ | async)"
+        src="{{image.url}}" 
+        daff-gallery-image
+        [selected]="image.url === (selectedImage$ | async)"
         (click)="select(image.url)">
     </div>
   </div>
