--- libs/design/src/molecules/gallery-image/gallery-image.component.html
@@ -1,8 +1,0 @@
-<div class="image-gallery__gallery-image-wrapper">
-  <img 
-    alt="{{image.label}}"
-    src="{{image.url}}"  
-    class="image-gallery__gallery-image"
-    [class.image-gallery__gallery-image--selected]="selected"
-    (click)="select(image)">
-</div>
