--- libs/design/src/molecules/gallery-image/gallery-image.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { DaffGalleryImageComponent } from './gallery-image.component';
-
-@NgModule({
-  imports: [
-    CommonModule
-  ],
-  declarations: [
-    DaffGalleryImageComponent
-  ],
-  exports: [
-    DaffGalleryImageComponent
-  ]
-})
-export class DaffGalleryImageModule { }
