--- libs/design/src/molecules/gallery-image/public_api.ts
@@ -1,2 +1,0 @@
-export { DaffGalleryImageComponent } from './gallery-image.component';
-export { DaffGalleryImageModule } from './gallery-image.module';
