+++ libs/design/src/molecules/image-gallery/image-gallery/image-gallery.component.html
