+++ libs/design/src/molecules/image-gallery/image-gallery/image-gallery.component.scss
@@ -1,4 +1,4 @@
-@import '../../scss/daff-util';
+@import 'daff-util';
 
 .image-gallery {
   display: flex;
