+++ libs/design/src/molecules/image-gallery/public_api.ts
@@ -1,2 +1,4 @@
-export { ImageGalleryComponent } from './image-gallery.component';
+export { ImageGalleryComponent } from './image-gallery/image-gallery.component';
+export { GalleryImageComponent } from './gallery-image/gallery-image.component';
+export { SelectedGalleryImageComponent } from './selected-gallery-image/selected-gallery-image.component';
 export { DaffImageGalleryModule } from './image-gallery.module';
