+++ libs/design/src/molecules/image-gallery/selected-gallery-image/selected-gallery-image.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: '[daff-selected-gallery-image]',
+  template: '<ng-content></ng-content>',
+  host: { class: 'image-gallery__gallery-image--selected' }
+})
+export class SelectedGalleryImageComponent {}
