+++ libs/design/src/public_api.ts
@@ -11,7 +11,6 @@ export * from './atoms/container/public_api';
 export * from './molecules/accordion/public_api';
 export * from './molecules/backdrop/public_api';
 export * from './molecules/button-set/public_api';
-export * from './molecules/gallery-image/public_api';
 export * from './molecules/image-gallery/public_api';
 export * from './molecules/image-list/public_api';
 export * from './molecules/input-validator/public_api';
